Create Simple_bs4_Web_Crawler
As of 11/10/2020, This simple program just pulls all URLs found in the HTML of the website you provide. It was created in Thonny but I would prefer to make this an actual program.

# Things to add
Actual "crawling" feature
Search for vital websites ( index, robots, login pages, parent directories, etc. etc. )
Add a GUI and output to a file